backend/requirements.txt
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2024.1
requests>=2.31.0
feedparser>=6.0.0
groq>=0.4.0
edge-tts>=6.1.0
